+++ src/main/java/de/maxhenkel/inhabitor/mixin/ChunkMapMixin.java
@@ -1,25 +1,25 @@
 package de.maxhenkel.inhabitor.mixin;
 
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import de.maxhenkel.inhabitor.Inhabitor;
 import net.minecraft.server.level.ChunkMap;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import java.util.function.BooleanSupplier;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.Executor;
 
 @Mixin(ChunkMap.class)
 public class ChunkMapMixin {
 
-    @Inject(method = "processUnloads", at = @At("HEAD"))
-    private void setTickSaveState(BooleanSupplier shouldKeepTicking, CallbackInfo ci) {
-        Inhabitor.IS_TICK_SAVE.set(true);
-    }
-
-    @Inject(method = "processUnloads", at = @At("RETURN"))
-    private void unsetTickSaveState(BooleanSupplier shouldKeepTicking, CallbackInfo ci) {
-        Inhabitor.IS_TICK_SAVE.set(false);
+    @WrapOperation(method = "scheduleUnload", at = @At(value = "INVOKE", target = "Ljava/util/concurrent/CompletableFuture;thenRunAsync(Ljava/lang/Runnable;Ljava/util/concurrent/Executor;)Ljava/util/concurrent/CompletableFuture;"))
+    private CompletableFuture<Void> scheduleUnload(CompletableFuture instance, Runnable action, Executor executor, Operation<CompletableFuture<Void>> original) {
+        return original.call(instance, (Runnable) () -> {
+            Inhabitor.IS_SCHEDULE_UNLOAD.set(true);
+            action.run();
+            Inhabitor.IS_SCHEDULE_UNLOAD.set(false);
+        }, executor);
     }
 
 }