+++ src/main/java/de/maxhenkel/inhabitor/config/ModConfig.java
@@ -8,7 +8,7 @@ public class ModConfig {
     public ConfigEntry<Integer> inhabitorCommandPermissionLevel;
 
     public ModConfig(ConfigBuilder builder) {
-        inhabitorCommandPermissionLevel = builder.integerEntry("inhabitor_command_permission_level", 2, 0, 16);
+        inhabitorCommandPermissionLevel = builder.integerEntry("inhabitor_command_permission_level", 2, 0, 4);
     }
 
 }