+++ changelog.md
@@ -1,3 +1,1 @@
-- Fixed set command adding time instead of setting
-- Added get command
-- Fixed inhabited time fix not working properly anymore
+- Updated to 1.21.11